Identified an issue when saving a job after editing it.
In the GetJobs method, when instantiating the Job class, temporaryJobID was being used instead of the first column of the row that was being read from the database table Job.

As temporaryJobID wasn't being set anywhere other than outside of the while loop, the number -1 was being set as the JobID for each Job instantiated. As a result, when trying to update one of those Jobs, the Jobs list was being indexed with -1 (the JobID of that Job) which was outside the bounds of the BindableList.

I may need to make additional modifications so that Jobs in the Jobs list is found specifically by its JobID.